Move components and services into respective folders
