+++ src/frontend/src/app/components/layout/header/header.component.ts
