+++ src/frontend/src/app/components/register/register.component.scss
