+++ src/frontend/src/app/components/register/register.component.spec.ts
