+++ src/frontend/src/app/services/guards/auth-guard.service.ts
